Refactoring of GLIR queues: 1 on canvas and 1 on context
All objects that can be shared accross canvases push their
commands to the context queue. All other commands go into
the canvas queue.

Also gloo objects start out with a temporary private queue,
which gets replaced when they are associated with a canvas.
This happens when on Program.draw and FrameBuffer.activate.